fix: clean up MCP auditability
