+++ internal/service/frontend/server.go
@@ -28,7 +28,6 @@ import (
 	"github.com/go-chi/chi/v5/middleware"
 	"github.com/go-chi/cors"
 	"github.com/go-chi/httplog/v2"
-	"github.com/google/uuid"
 	"github.com/prometheus/client_golang/prometheus"
 
 	"github.com/dagucloud/dagu/internal/agent"
@@ -1377,165 +1376,6 @@ func (srv *Server) setupMCPRoute(ctx context.Context, r *chi.Mux) {
 	})
 
 	logger.Info(ctx, "MCP route configured", slog.String("path", mcpPath))
-}
-
-func (srv *Server) mcpAuditSeedMiddleware() func(http.Handler) http.Handler {
-	return func(next http.Handler) http.Handler {
-		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			source := &audit.SourceContext{
-				Source:        "mcp",
-				Surface:       string(authmodel.APIKeySurfaceMCP),
-				RequestID:     uuid.NewString(),
-				CorrelationID: uuid.NewString(),
-				Transport:     "streamable_http",
-			}
-			if requestedWorkspace := strings.TrimSpace(r.URL.Query().Get("workspace")); requestedWorkspace != "" {
-				source.RequestedWorkspace = requestedWorkspace
-				source.ResolvedWorkspace = canonicalMCPWorkspace(requestedWorkspace)
-			}
-			next.ServeHTTP(w, r.WithContext(audit.WithSourceContext(r.Context(), source)))
-		})
-	}
-}
-
-func (srv *Server) mcpAuditSubjectMiddleware() func(http.Handler) http.Handler {
-	return func(next http.Handler) http.Handler {
-		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			ctx := withMCPAuditCredentialContext(r.Context(), nil)
-			if apiKey, ok := authmodel.APIKeyFromContext(ctx); ok {
-				if user, ok := mcpAuditUserForAPIKey(apiKey); ok {
-					ctx = authmodel.WithUser(ctx, user)
-				}
-			}
-			next.ServeHTTP(w, r.WithContext(ctx))
-		})
-	}
-}
-
-func (srv *Server) logMCPAuthDenied(r *http.Request, reason string, apiKey *authmodel.APIKey) {
-	if srv == nil || srv.apiV1 == nil || r == nil {
-		return
-	}
-	ctx := withMCPAuditCredentialContext(r.Context(), apiKey)
-	if user, ok := mcpAuditUserForAPIKey(apiKey); ok {
-		ctx = authmodel.WithUser(ctx, user)
-	}
-	details := map[string]any{
-		"result":        "denied",
-		"denial_reason": reason,
-		"resource_type": "mcp_request",
-		"resource_id":   r.URL.Path,
-		"http_method":   r.Method,
-	}
-	srv.apiV1.LogAudit(ctx, audit.CategoryMCP, "mcp.request.denied", details)
-}
-
-func withMCPAuditCredentialContext(ctx context.Context, deniedAPIKey *authmodel.APIKey) context.Context {
-	source, ok := audit.SourceContextFromContext(ctx)
-	if !ok || source == nil {
-		source = &audit.SourceContext{
-			Source:        "mcp",
-			Surface:       string(authmodel.APIKeySurfaceMCP),
-			RequestID:     uuid.NewString(),
-			CorrelationID: uuid.NewString(),
-			Transport:     "streamable_http",
-		}
-	}
-	if source.Source == "" {
-		source.Source = "mcp"
-	}
-	if source.Surface == "" {
-		source.Surface = string(authmodel.APIKeySurfaceMCP)
-	}
-	if source.Transport == "" {
-		source.Transport = "streamable_http"
-	}
-
-	if user, ok := authmodel.UserFromContext(ctx); ok && user != nil {
-		source.SubjectID = user.ID
-		source.SubjectName = user.Username
-		source.SubjectType = "user"
-	}
-
-	apiKey := deniedAPIKey
-	if apiKey == nil {
-		apiKey, _ = authmodel.APIKeyFromContext(ctx)
-	}
-	if apiKey != nil {
-		apiKey = authmodel.NormalizeAPIKeyMetadata(apiKey)
-		source.CredentialID = apiKey.ID
-		source.CredentialName = apiKey.Name
-		source.CredentialType = "api_key"
-		source.CredentialAllowedSurfaces = apiKeySurfaceStrings(apiKey.AllowedSurfaces)
-		source.AttributionClass = string(apiKey.AttributionClass)
-		switch apiKey.AttributionClass {
-		case authmodel.APIKeyAttributionUserOwned:
-			source.SubjectType = "user"
-			source.SubjectID = apiKey.OwnerUserID
-			source.SubjectName = apiKey.OwnerUsername
-			source.CredentialOwnerID = apiKey.OwnerUserID
-		case authmodel.APIKeyAttributionServiceAccount:
-			source.SubjectType = "service_account"
-			source.SubjectID = apiKey.ServiceAccountID
-			source.SubjectName = apiKey.ServiceAccountName
-			source.ServiceAccountID = apiKey.ServiceAccountID
-		}
-	} else if source.CredentialType == "" {
-		source.CredentialType = credentialTypeFromContext(ctx)
-	}
-
-	return audit.WithSourceContext(ctx, source)
-}
-
-func mcpAuditUserForAPIKey(apiKey *authmodel.APIKey) (*authmodel.User, bool) {
-	if apiKey == nil {
-		return nil, false
-	}
-	apiKey = authmodel.NormalizeAPIKeyMetadata(apiKey)
-	switch apiKey.AttributionClass {
-	case authmodel.APIKeyAttributionUserOwned:
-		if apiKey.OwnerUserID == "" {
-			return nil, false
-		}
-		return &authmodel.User{
-			ID:              apiKey.OwnerUserID,
-			Username:        apiKey.OwnerUsername,
-			Role:            apiKey.Role,
-			WorkspaceAccess: authmodel.CloneWorkspaceAccess(apiKey.WorkspaceAccess),
-		}, true
-	case authmodel.APIKeyAttributionServiceAccount:
-		return &authmodel.User{
-			ID:              apiKey.ServiceAccountID,
-			Username:        apiKey.ServiceAccountName,
-			Role:            apiKey.Role,
-			WorkspaceAccess: authmodel.CloneWorkspaceAccess(apiKey.WorkspaceAccess),
-		}, true
-	default:
-		return nil, false
-	}
-}
-
-func credentialTypeFromContext(ctx context.Context) string {
-	if _, ok := authmodel.UserFromContext(ctx); ok {
-		return "session"
-	}
-	return "none"
-}
-
-func apiKeySurfaceStrings(surfaces []authmodel.APIKeySurface) []string {
-	normalized := authmodel.NormalizeAPIKeySurfaces(surfaces)
-	out := make([]string, 0, len(normalized))
-	for _, surface := range normalized {
-		out = append(out, string(surface))
-	}
-	return out
-}
-
-func canonicalMCPWorkspace(workspace string) string {
-	if workspace == "" {
-		return "default"
-	}
-	return workspace
 }
 
 func clearWriteDeadlineMiddleware(next http.Handler) http.Handler {
